bootstap app in standalone manner
