+++ src/app/app.component.spec.ts
@@ -4,8 +4,8 @@ import { AppComponent } from './app.component';
 describe('AppComponent', () => {
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [AppComponent],
-    }).compileComponents();
+    imports: [AppComponent],
+}).compileComponents();
   });
 
   it('should create the app', () => {
