+++ src/app/http-error.interceptor.ts
@@ -1,25 +1,22 @@
-import { HttpErrorResponse, HttpEvent, HttpHandler, HttpInterceptor, HttpRequest } from '@angular/common/http';
-import { Injectable } from '@angular/core';
+import { HttpErrorResponse, HttpInterceptorFn } from '@angular/common/http';
+import { inject } from '@angular/core';
 import { MatSnackBar } from '@angular/material/snack-bar';
-import { Observable, throwError } from 'rxjs';
+import { throwError } from 'rxjs';
 import { catchError } from 'rxjs/operators';
 
-@Injectable()
-export class HttpErrorInterceptor implements HttpInterceptor {
-  constructor(private snackbar: MatSnackBar) {}
+export const httpErrorInterceptor: HttpInterceptorFn = (req, next) => {
+  const snackbar = inject(MatSnackBar);
 
-  intercept(request: HttpRequest<unknown>, next: HttpHandler): Observable<HttpEvent<unknown>> {
-    return next.handle(request).pipe(
-      catchError(err => {
-        let message: string;
-        if (err instanceof HttpErrorResponse) {
-          message = `Http error, status ${err.status}, url ${err.url}: ${err.message}`;
-        } else {
-          message = `Unknown error: ${err?.message || err?.error || err}`;
-        }
-        this.snackbar.open(message, 'Close', { duration: 15000, panelClass: 'error-snackbar' });
-        return throwError(() => err);
-      }),
-    );
-  }
-}
+  return next(req).pipe(
+    catchError(err => {
+      let message: string;
+      if (err instanceof HttpErrorResponse) {
+        message = `Http error, status ${err.status}, url ${err.url}: ${err.message}`;
+      } else {
+        message = `Unknown error: ${err?.message || err?.error || err}`;
+      }
+      snackbar.open(message, 'Close', { duration: 15000, panelClass: 'error-snackbar' });
+      return throwError(() => err);
+    }),
+  );
+};
